Fix a bug that would align atlasTS to TS, which should be vice versa.
It was in line 357 of funcnorm.m.

TS should be aligned to atlasTS (see line 414 of funcnorm.m, and lines 841 and
971 of connnorm.m), and then warped to be used by the new atlas.

The original code would get the warp by aligning atlasTS to TS, and then use
the warp of atlasTS to warp TS, which would make TS more deviated from the
atlas during the 0th pass.